Put portfolios in the db

+++ routes/co-app.js
@@ -4,6 +4,8 @@ const md5 = require('md5');
 const uuidv4 = require('uuid/v4');
 const nodemailer = require('nodemailer');
 
+const Portfolio = require('./portfolio')
+
 let transporter;
 router.use((req, res, next) => {
   transporter = nodemailer.createTransport({
@@ -14,42 +16,70 @@ router.use((req, res, next) => {
   next();
 });
 
-// TODO: in the future, validate sessions on every request if I want to allow them to timeout
-/*
-router.use((req, res, next) => {
-  const session = req.get('session'),
-        username = req.get('username');
+router.post('/session', (req, res, next) => {
+  const values = req.body,
+        username = values.username,
+        password = values.password,
+        hash = md5(password);
   return req.client.query({
-    text: 'select current_session from users where username = $1::text',
+    text: 'select * from users where username = $1::text',
+    values: [username]
+  }).then((results) => {
+    const user = results.rows[0],
+          currentSession = user.current_session,
+          passwordHash = user.hashed_password;
+    if (passwordHash === hash) {
+      if (currentSession) {
+        return res.json(currentSession);
+      }
+      const session = uuidv4();
+      return req.client.query({
+        text: 'update users set current_session = $1::text where id = $2::bigint',
+        values: [session, user.id]
+      }).then(() => {
+        req.client.release();
+        return res.json(session);
+      });
+    } else {
+      req.client.release();
+      return res.sendStatus(403);
+    }
+  })
+});
+
+router.use((req, res, next) => {
+  const session = req.get('x-session-id'),
+        username = req.get('x-username');
+  return req.client.query({
+    text: 'select * from users where username = $1::text',
     values: [username]
   }).then((results) => {
     const user = results.rows[0],
           currentSession = user.current_session;
     if (currentSession === session) {
+      req.userId = user.id;
+      const portfolio = new Portfolio(req.client, user.id);
+      req.app.set('portfolio', portfolio);
       return next();
     } else {
       return res.sendStatus(401);
     }
-  })
+  });
 });
-*/
 
-/*
-router.get('/', function(req, res, next) {
-  res.json({ message: 'hi there' });
-});
-*/
+const pc = require('./pc-routes');
+router.use('/portfolio', pc);
 
-const getMessagesFromJobID = (client, jobID ) => {
+const getMessagesFromJobID = (client, userId, jobID) => {
   return client.query({
-  	text: 'select * from messages where job_id = $1::bigint',
-  	values: [jobID]
+  	text: 'select * from messages where user_id = $1::bigint and job_id = $2::bigint',
+  	values: [userId, jobID]
   });
 };
 
-const getFactsFromJobID = (client, jobID) => {
+const getFactsFromJobID = (client, userId, jobID) => {
   return client.query({
-    text: 'select * from facts where job_id = $1::bigint',
+    text: 'select * from facts where user_id = $1::bigint and job_id = $2::bigint',
     values: [jobID]
   });
 };
@@ -58,20 +88,22 @@ router.post('/jobs', (req, res, next) => {
   const job = req.body,
         email = job.email;
   return req.client.query({
-    text: 'insert into jobs (email, user_id) values($1::text, $2::bigint) returning *', 
-    values: [email, 1] // TODO: pass in other users' IDs
+    text: 'insert into jobs (email, user_id) values($1::text, $2::bigint) returning *',
+    values: [email, req.userId]
   }).then((result) => {
     const newJob = result.rows[0];
-    req.client.end();
+    req.client.release();
     return res.json(newJob);
   });
 });
 
 router.get('/jobs', function(req, res, next) {
   return req.client.query({
-    text: `select j.*, m.latest_msg, f.key, f.value from jobs j 
+    text: `select j.*, m.latest_msg, f.key, f.value from jobs j
       left outer join (select job_id, max(datetime) as latest_msg from messages group by job_id) m on m.job_id=j.id
-      left outer join facts f on f.job_id=j.id`
+      left outer join facts f on f.job_id=j.id
+      where j.user_id = $1::bigint and m.user_id = $1::bigint and f.user_id = $1::bigint`,
+    values: [req.userId]
   }).then((results) => {
     const jobObj = {};
     results.rows.forEach((job) => {
@@ -90,7 +122,7 @@ router.get('/jobs', function(req, res, next) {
       }
     });
     const jobs = Object.keys(jobObj).map((key) => jobObj[key]);
-    req.client.end();
+    req.client.release();
     return res.json(jobs);
   });
 });
@@ -98,10 +130,10 @@ router.get('/jobs', function(req, res, next) {
 router.put('/jobs/:job_id', (req, res, next) => {
   const job = req.body;
   return req.client.query({
-    text: 'update jobs set company = $1::text, archived = $2::boolean where id=$3::bigint',
-    values: [job.company, job.archived, req.params.job_id]
+    text: 'update jobs set company = $1::text, archived = $2::boolean where user_id = $3::bigint and id=$4::bigint',
+    values: [job.company, job.archived, req.userId, req.params.job_id]
   }).then(() => {
-    req.client.end();
+    req.client.release();
     res.sendStatus(200);
   });
 });
@@ -109,23 +141,25 @@ router.put('/jobs/:job_id', (req, res, next) => {
 router.delete('/jobs/:job_id', (req, res, next) => {
   const jobID = req.params.job_id;
   return req.client.query({
-    text: 'delete from messages where job_id = $1::bigint',
-    values: [jobID]
+    text: 'delete from messages where user_id = $1::bigint job_id = $2::bigint',
+    values: [req.userId, jobID]
   }).then(() => {
     return req.client.query({
-      text: 'delete from facts where job_id = $1::bigint',
-      values: [jobID]
+      text: 'delete from facts where user_id = $1::bigint and job_id = $2::bigint',
+      values: [req.userId, jobID]
     });
   }).then(() => {
     return req.client.query({
-      text: 'delete from jobs where id = $1::bigint',
-      values: [jobID]
+      text: 'delete from jobs where user_id = $1::bigint and id = $1::bigint',
+      values: [req.userId, jobID]
     });
   }).then(() => {
-    // req.client.end() // throws an error?
+    req.client.release();
     return res.sendStatus(200);
   });
 });
+
+// TODO: add user_id to the queries below to make them more secure
 
 router.post('/jobs/:job_id/messages', (req, res, next) => {
   const type = 'text',
@@ -157,9 +191,11 @@ router.post('/jobs/:job_id/messages', (req, res, next) => {
         subject: 'New message from ' + msg.sender,
         text: `${msg.value}\nView discussion: http://counteroffer.me/${username}/#!/contact?job=${jobID}`
       }).then(() => {
+        req.client.release();
         return res.json(msg);
       });
     } else {
+      req.client.release();
       return res.sendStatus(200);
     }
   });
@@ -167,14 +203,14 @@ router.post('/jobs/:job_id/messages', (req, res, next) => {
 
 router.get('/jobs/:job_id/messages', (req, res, next) => {
 	return getMessagesFromJobID(req.client, req.params.job_id).then((results) => {
-    req.client.end();
+    req.client.release();
     return res.json(results.rows);
   });
 });
 
 router.get('/jobs/:job_id/facts', (req, res, next) => {
   return getFactsFromJobID(req.client, req.params.job_id).then((results) => {
-    req.client.end();
+    req.client.release()
     return res.json(results.rows);
   });
 });
@@ -188,7 +224,7 @@ router.post('/jobs/:job_id/facts', (req, res, next) => {
     text: 'insert into facts (key, value, job_id) values ($1::text, $2::text, $3::bigint) returning *',
     values: [key, value, jobID]
   }).then((results) => {
-    req.client.end();
+    req.client.release()
     return results.rows && results.rows.length === 1 ? res.json(results.rows)[0] : null;
   });
 });
@@ -199,7 +235,7 @@ router.put('/jobs/:job_id/facts/:fact_id', (req, res, next) => {
     text: 'update facts set key = $1::text, value = $2::text where id = $3::bigint',
     values: [fact.key, fact.value, fact.id]
   }).then(() => {
-    req.client.end();
+    req.client.release()
     return res.sendStatus(200);
   });
 });
@@ -209,37 +245,8 @@ router.delete('/jobs/:job_id/facts/:fact_id', (req, res, next) => {
     text: 'delete from facts where id=$1::bigint and job_id=$2::bigint',
     values: [req.params.fact_id, req.params.job_id]
   }).then(() => {
-    req.client.end();
+    req.client.release()
     return res.sendStatus(200);
-  })
-});
-
-router.post('/session', (req, res, next) => {
-  const values = req.body,
-        username = values.username,
-        password = values.password,
-        hash = md5(password);
-  return req.client.query({
-    text: 'select * from users where username = $1::text',
-    values: [username]
-  }).then((results) => {
-    const user = results.rows[0],
-          currentSession = user.current_session,
-          passwordHash = user.hashed_password;
-    if (passwordHash === hash) {
-      if (currentSession) {
-        return res.json(currentSession);
-      }
-      const session = uuidv4();
-      return req.client.query({
-        text: 'update users set current_session = $1::text where id = $2::bigint',
-        values: [session, user.id]
-      }).then(() => {
-        return res.json(session);
-      });
-    } else {
-      return res.sendStatus(403);
-    }
   })
 });
